Create swift file for NSObject extension
