+++ VendingMachine/VendingMachine/Drink/Drink.swift
@@ -70,31 +70,6 @@ extension Drink: Comparable {
     }
 }
 
-extension NSObject {
-    var className: String {
-        return String(describing: type(of: self)).components(separatedBy: ".").last!
-    }
-    
-    class var className: String {
-        return String(describing: self).components(separatedBy: ".").last!
-    }
-    
-    var typeName: String {
-        var typeName = ""
-        switch self {
-        case is Milk:
-            typeName = Milk.className()
-        case is SoftDrink:
-            typeName = SoftDrink.className()
-        case is Coffee:
-            typeName  = Coffee.className()
-        default:
-            typeName = self.className
-        }
-        return typeName
-    }
-}
-
 let dateFormatter: DateFormatter = {
     let formatter = DateFormatter()
     formatter.dateFormat = "yyyyMMdd"
